fix formatting and move code
