+++ msvc/mono.sln
@@ -225,19 +225,25 @@ Global
 		Debug|ARM64 = Debug|ARM64
 		Debug|Win32 = Debug|Win32
 		Debug|x64 = Debug|x64
+		Debug-target-arm64|ARM64 = Debug-target-arm64|ARM64
+		Debug-target-arm64|Win32 = Debug-target-arm64|Win32
+		Debug-target-arm64|x64 = Debug-target-arm64|x64
 		Release|ARM64 = Release|ARM64
 		Release|Win32 = Release|Win32
 		Release|x64 = Release|x64
+		Release-target-arm64|ARM64 = Release-target-arm64|ARM64
+		Release-target-arm64|Win32 = Release-target-arm64|Win32
+		Release-target-arm64|x64 = Release-target-arm64|x64
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B7098DFA-31E6-4006-8A15-1C9A4E925149}.Debug|ARM64.ActiveCfg = Debug|x64
-		{B7098DFA-31E6-4006-8A15-1C9A4E925149}.Debug|ARM64.Build.0 = Debug|x64
+		{B7098DFA-31E6-4006-8A15-1C9A4E925149}.Debug|ARM64.ActiveCfg = Debug-target-arm64|x64
+		{B7098DFA-31E6-4006-8A15-1C9A4E925149}.Debug|ARM64.Build.0 = Debug-target-arm64|x64
 		{B7098DFA-31E6-4006-8A15-1C9A4E925149}.Debug|Win32.ActiveCfg = Debug|Win32
 		{B7098DFA-31E6-4006-8A15-1C9A4E925149}.Debug|Win32.Build.0 = Debug|Win32
 		{B7098DFA-31E6-4006-8A15-1C9A4E925149}.Debug|x64.ActiveCfg = Debug|x64
 		{B7098DFA-31E6-4006-8A15-1C9A4E925149}.Debug|x64.Build.0 = Debug|x64
-		{B7098DFA-31E6-4006-8A15-1C9A4E925149}.Release|ARM64.ActiveCfg = Release|x64
-		{B7098DFA-31E6-4006-8A15-1C9A4E925149}.Release|ARM64.Build.0 = Release|x64
+		{B7098DFA-31E6-4006-8A15-1C9A4E925149}.Release|ARM64.ActiveCfg = Release-target-arm64|x64
+		{B7098DFA-31E6-4006-8A15-1C9A4E925149}.Release|ARM64.Build.0 = Release-target-arm64|x64
 		{B7098DFA-31E6-4006-8A15-1C9A4E925149}.Release|Win32.ActiveCfg = Release|Win32
 		{B7098DFA-31E6-4006-8A15-1C9A4E925149}.Release|Win32.Build.0 = Release|Win32
 		{B7098DFA-31E6-4006-8A15-1C9A4E925149}.Release|x64.ActiveCfg = Release|x64
